requirements.txt
tensorflow~=2.16.1
keras~=3.3.2
typing_extensions~=4.11.0
pillow~=10.2.0
pip~=23.3.2
wheel~=0.43.0
gast~=0.5.4
numpy~=1.26.4
h5py~=3.11.0
pytest~=8.2.2
pytz~=2024.1
rich~=13.7.1
Pygments~=2.17.2
optree~=0.11.0
packaging~=24.0
scipy~=1.14.0
pandas~=2.2.2
requests~=2.31.0
namex~=0.0.8
mdurl~=0.1.2
setuptools~=69.2.0
six~=1.16.0
threadpoolctl~=3.5.0
matplotlib~=3.8.4
joblib~=1.4.2
MarkupSafe~=2.1.5
python-dateutil~=2.9.0.post0
certifi~=2024.2.2
seaborn~=0.13.2
cycler~=0.12.1
urllib3~=2.2.1
Markdown~=3.6
idna~=3.7
contourpy~=1.2.1
fonttools~=4.51.0
pyparsing~=3.1.2
termcolor~=2.4.0
kiwisolver~=1.4.5
flatbuffers~=24.3.25
wrapt~=1.16.0
astunparse~=1.6.3
tensorboard~=2.16.2
Werkzeug~=3.0.2
scikit-learn~=1.5.0